requirements.txt
fastapi
uvicorn
python-dotenv
boto3
firebase-admin
